重複処理を削除（リファクタリング）

+++ app/controllers/items_controller.rb
@@ -12,15 +12,11 @@ class ItemsController < ApplicationController
 
     # 「チェックしたアイテム」機能
     checked_item = current_user.checked_items.find_by(item_id: @item.id)
-    if checked_item == nil
-      # 同じアイテムを過去に閲覧していなければレコード作成
-      current_user.checked_items.create(item_id: @item.id)
-    else
-      # 閲覧していれば既存のレコードを削除
-      checked_item.destroy
-      current_user.checked_items.create(item_id: @item.id)
-    end
+    # 既にアイテムを閲覧していれば既存のレコードを削除
+    checked_item.destroy if checked_item
+    current_user.checked_items.create(item_id: @item.id)
 
+    # チェックしたアイテムが規定数を超えた場合、古いレコードを削除
     if current_user.checked_items.length > 20
       checked_items.first.destroy
     end
